chore: redefine exports
